Added viewport meta tags to both dialogs
These tags, along with the ones in the initial commit will be used to
enable mobile viewing. Based on my desktop, I am able to easily view the
changes of adding these tags, as the query tags take effect on every web
page. Before, it was just the mozilla web page. Which, by the way, is
working better as well.